# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SDL DB Backup API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/restore/validate": {
      "post": {
        "summary": "Validate a logical backup run without restoring it",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["run_id"],
                "properties": {
                  "run_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result"
          }
        }
      }
    },
    "/api/v1/restore/test": {
      "post": {
        "summary": "Run a full sandbox restore test for a logical backup run",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["run_id"],
                "properties": {
                  "run_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sandbox restore validation result"
          }
        }
      }
    }
  }
}